+++ changelog.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2018-12-24
+
+### Added
+
+- Added a new type `JWTGenerateOptions`.
+
+### Changed
+
+- Renamed `add` to `generate` to better describe what it does. And changed its
+  signature to (only) accept options of type `JWTGenerateOptions`.
+
+- Allow alternate audience when verifying by providing an additional argument to
+  the method.
+
+- Made `verify`, `verifyHeader` and `decode` accept `undefined` as a value for
+  token/header.
+
+- Updated _some_ examples in readme.
+
 ## [0.1.5] - 2018-12-23
 
 ### Added
@@ -71,7 +90,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial release
 
-[Unreleased]: https://github.com/revam/node-jwt-manager/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/revam/node-jwt-manager/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/revam/node-jwt-manager/compare/v0.1.5...v0.2.0
 [0.1.5]: https://github.com/revam/node-jwt-manager/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/revam/node-jwt-manager/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/revam/node-jwt-manager/compare/v0.1.2...v0.1.3